+++ src/main/java/KTPM/example/BlueMoon/Iservice/Inhankhau.java
@@ -1,4 +1,5 @@
 package KTPM.example.BlueMoon.Iservice;
 
 public interface Inhankhau {
+
 }
