+++ src/main/java/KTPM/example/BlueMoon/repository/hokhauRepository.java
@@ -4,6 +4,12 @@ import KTPM.example.BlueMoon.model.Hokhau;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 @Repository
 public interface hokhauRepository extends JpaRepository<Hokhau, Integer> {
+    Hokhau findByHokhauid(int id);
+
+
+
+
 }
