+++ src/main/java/KTPM/example/BlueMoon/repository/nhankhauRepository.java
@@ -4,6 +4,9 @@ import KTPM.example.BlueMoon.model.Nhankhau;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface nhankhauRepository extends JpaRepository<Nhankhau, Integer> {
+
 }
