+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ server.port= 8081
 
 spring.datasource.url=jdbc:postgresql://localhost:5432/Bluemoon
 spring.datasource.username=postgres
-spring.datasource.password=1
+spring.datasource.password=1111
 
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=update
