--- src/src/main/java/KTPM/example/BlueMoon/BlueMoonApplication.java
@@ -1,15 +1,0 @@
-package KTPM.example.BlueMoon;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BlueMoonApplication {
-
-    public static void main(String[] args) {
-
-		SpringApplication.run(BlueMoonApplication.class, args);
-		System.out.println("Hello Blue Moon");
-    }
-
-}
