--- src/src/main/java/KTPM/example/BlueMoon/Iservice/Iaccount.java
@@ -1,5 +1,0 @@
-package KTPM.example.BlueMoon.Iservice;
-
-public interface Iaccount {
-    boolean login(String username, String password);
-}
