--- src/src/main/java/KTPM/example/BlueMoon/Iservice/Inhankhau.java
@@ -1,5 +1,0 @@
-package KTPM.example.BlueMoon.Iservice;
-
-public interface Inhankhau {
-
-}
