--- src/src/main/java/KTPM/example/BlueMoon/Iservice/Ipay.java
@@ -1,4 +1,0 @@
-package KTPM.example.BlueMoon.Iservice;
-
-public interface Ipay {
-}
