--- src/src/main/java/KTPM/example/BlueMoon/Iservice/Iuser.java
@@ -1,16 +1,0 @@
-package KTPM.example.BlueMoon.Iservice;
-
-import java.util.List;
-import java.util.Map;
-
-public interface Iuser {
-    boolean login(String username, String password);
-
-    boolean register(String username, String password, int roleid);
-
-    boolean changePassword(int userid, String password);
-
-    boolean deleteUser(String username);
-
-    List<Map<String,Object>> getAccounts();
-}
