--- src/src/main/java/KTPM/example/BlueMoon/config/SecurityConfig.java
@@ -1,55 +1,0 @@
-package KTPM.example.BlueMoon.config;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.List;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                // Tắt form login mặc định
-                .formLogin(AbstractHttpConfigurer::disable)
-                // Tắt chuyển hướng đến trang login
-                .httpBasic(AbstractHttpConfigurer::disable)
-                // Cấu hình CORS
-                .cors(cors -> cors.configurationSource(corsConfigurationSource()))
-                // Tắt CSRF cho API (hoặc cấu hình phù hợp)
-                .csrf(AbstractHttpConfigurer::disable)
-                // Phân quyền request
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/**").permitAll() // Cho phép truy cập không cần auth
-                        .anyRequest().authenticated() // Các request khác cần auth
-                )
-                // Cấu hình session stateless (phù hợp cho API)
-                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-
-        return http.build();
-    }
-
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration configuration = new CorsConfiguration();
-        configuration.setAllowedOrigins(List.of("http://localhost:3000")); // Frontend URL
-        configuration.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
-        configuration.setAllowedHeaders(List.of("*"));
-        configuration.setAllowCredentials(true); // Cho phép gửi cookie/session
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-    }
-}
