--- src/src/main/java/KTPM/example/BlueMoon/config/WebConfig.java
@@ -1,17 +1,0 @@
-package KTPM.example.BlueMoon.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("http://localhost:3000") // Cho phép Frontend truy cập
-                .allowCredentials(true) // Cho phép gửi cookie (session)
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
-                .allowedHeaders("*");
-    }
-}
