--- src/src/main/java/KTPM/example/BlueMoon/controller/totruongController.java
@@ -1,11 +1,0 @@
-package KTPM.example.BlueMoon.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/totruong")
-@RequiredArgsConstructor
-public class totruongController {
-}
