--- src/src/main/java/KTPM/example/BlueMoon/model/Role.java
@@ -1,26 +1,0 @@
-package KTPM.example.BlueMoon.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Entity
-@Table(name = "roles")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int roleid;
-
-    private String rolename;
-
-    @OneToMany(mappedBy = "role")
-    private List<User> users;
-}
