--- src/src/main/java/KTPM/example/BlueMoon/model/User.java
@@ -1,28 +1,0 @@
-package KTPM.example.BlueMoon.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "users")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    private String username;
-    private String password;
-
-    @ManyToOne
-    @JoinColumn(name = "roleid")
-    private Role role;
-
-    private boolean enabled;
-}
