--- src/src/main/java/KTPM/example/BlueMoon/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package KTPM.example.BlueMoon.repository;
-
-import KTPM.example.BlueMoon.model.Lichsuthaydoi;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Lichsuthaydoi, Integer> {
-}
