--- src/src/main/java/KTPM/example/BlueMoon/repository/tamtrutamvangRepository.java
@@ -1,9 +1,0 @@
-package KTPM.example.BlueMoon.repository;
-
-import KTPM.example.BlueMoon.model.Tamtrutamvang;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface tamtrutamvangRepository extends JpaRepository<Tamtrutamvang, Integer> {
-}
