--- src/src/main/java/KTPM/example/BlueMoon/service/memberService.java
@@ -1,10 +1,0 @@
-package KTPM.example.BlueMoon.service;
-
-import KTPM.example.BlueMoon.Iservice.Imember;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class memberService implements Imember {
-}
