--- src/src/main/java/KTPM/example/BlueMoon/service/nhankhauService.java
@@ -1,17 +1,0 @@
-package KTPM.example.BlueMoon.service;
-
-import KTPM.example.BlueMoon.Iservice.Inhankhau;
-import KTPM.example.BlueMoon.model.Hokhau;
-import KTPM.example.BlueMoon.model.Nhankhau;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import KTPM.example.BlueMoon.repository.nhankhauRepository;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class nhankhauService implements Inhankhau {
-
-}
