--- src/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-spring.application.name=BlueMoon
-server.port= 8081
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/Bluemoon
-spring.datasource.username=postgres
-spring.datasource.password=1111
-
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.hibernate.ddl-auto=update
-
