--- src/src/test/java/KTPM/example/BlueMoon/BlueMoonApplicationTests.java
@@ -1,13 +1,0 @@
-package KTPM.example.BlueMoon;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BlueMoonApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
